Update pagination and add total pages to view_events
This commit updates the pagination feature in view_events to display the correct total number of pages based on the limit set. It also adds a new variable, total_pages, to the pagination dictionary to keep track of the total number of pages. Additionally, a new component, pagination.html, has been added to the get_events template to display the pagination feature.